profile routing and structural changes (#861)
* remove unnecessary type

* adjust components to take in props instead of querying directly

* add new v2 profile page with userId slug; delete old profile pages

* add new profile shortcut gateway

* modify v1 profiles so they continue working

* simplify userId logic

* move logic to server side

* remove isEditable prop from achievement component
